+++ spec/ductwork/dsl/definition_builder/collapse_spec.rb
@@ -29,7 +29,12 @@ RSpec.describe Ductwork::DSL::DefinitionBuilder, "#collapse" do
       { to: %w[MySecondStep.1], type: :expand, klass: "MyFirstStep" }
     )
     expect(definition[:edges]["MySecondStep.1"]).to eq(
-      { to: %w[MyThirdStep.2], type: :collapse, klass: "MySecondStep" }
+      {
+        to: %w[MyThirdStep.2],
+        type: :collapse,
+        klass: "MySecondStep",
+        barrier_node: "MyFirstStep.0",
+      }
     )
     expect(definition[:edges]["MyThirdStep.2"]).to eq({ klass: "MyThirdStep" })
   end